requirements.txt
openai>=1.30.0
streamlit==1.31.0
python-dotenv==1.0.1